Moved the breakdown chart to the charts
