+++ src/Containers/JobExplorer/JobExplorerListRow.js
@@ -5,7 +5,7 @@ import { Tr, Td, ExpandableRowContent } from '@patternfly/react-table';
 
 import { formatDateTime, formatJobType } from '../../Utilities/helpers';
 import JobStatus from '../../Components/JobStatus';
-import Breakdown from '../../Components/Breakdown';
+import Breakdown from '../../Charts/Breakdown';
 import {
   global_palette_green_300,
   global_palette_black_400,
